add delete functionality

+++ client/src/components/book/PageViewBook.js
@@ -1,7 +1,7 @@
 import '../account/pageViews/pageViews.css';
 import { useSelector, useDispatch } from 'react-redux';
 import { removeReadingPlan, readingCompleted } from '../../state/actions';
-import { updateExistingPlan } from '../../functions/readingPlanFunctions';
+import { updateExistingPlan, deletePlanFromServer } from '../../functions/readingPlanFunctions';
 import { useState } from 'react';
 
 function PageViewBook (props) {
@@ -9,6 +9,8 @@ function PageViewBook (props) {
   const token = useSelector(state => state.loginReducer.token);
   const dispatch = useDispatch();
   const [updatingInProgress, setUpdatingInProgress] = useState(false)
+  const [deleteBookMode, setDeleteBookMode] = useState(false);
+
   try {
   
   const {book} = props;
@@ -35,8 +37,6 @@ function PageViewBook (props) {
     return {to: 'finished'}
   }
 
-  console.log(scheme)
-
   function updateThePlanScheme (update_day) {
     const updatedScheme = {};
     for(let i = 1; i <= Object.keys(scheme).length; i++) {
@@ -48,7 +48,7 @@ function PageViewBook (props) {
     return updatedScheme;
   }
 
-  async function completeReading (update_planId, update_day, update_book) {
+  async function completeReading (update_planId, update_day) {
     setUpdatingInProgress(true)
     try {
       const updatedScheme = updateThePlanScheme(update_day);
@@ -68,11 +68,21 @@ function PageViewBook (props) {
 
   const percetage = calculatePercentageComplete();
   const {day, date, from, to} = getNextItemInScheme();
-  const displayDate = new Date(date).toDateString()
-  const completeButtonClassNames = updatingInProgress ? "btn cb-btn submit-btn cb-complete-btn .btn-inSearchMode" : "btn cb-btn submit-btn cb-complete-btn";
+  const displayDate = new Date(date).toDateString();
+  const modalClassNames = deleteBookMode ? "deleteModal" : "deleteModal hideModal";
 
-  const handleDeleteBook = (book_id) => {
-    dispatch(removeReadingPlan(book_id))
+  const handleDeleteBook = async () => {
+    try {
+      const book_id = book.id;
+      const isPlanDeletedFromServer = await deletePlanFromServer(book_id, token);
+      if(isPlanDeletedFromServer) {
+        dispatch(removeReadingPlan(book_id));
+        return
+      }
+      throw Error;
+    } catch (error) {
+      window.alert(`It looks like there was an error that occured when trying to delete your plan.  Please logout and try again.  If the problem persists, please contact us via the 'About' button at the top of the page.`);
+    }
   } 
 
   return (
@@ -94,15 +104,33 @@ function PageViewBook (props) {
             <p className="cb-para bold">Next Up:</p>
             <p className="cb-para">{displayDate}</p>
             <p className="cb-para">Read from {book.measure}: <span className="bold">{from}</span> <br/>to the end of: <span className="bold">{to}</span></p>
-            <div className="btn cb-btn submit-btn cb-complete-btn" onClick={()=>completeReading(book.id, day, book)}>MARK COMPLETE</div>
+            {updatingInProgress ?
+              <div className="btn cb-btn submit-btn cb-complete-btn btn-inSearchMode" disabled>MARK COMPLETE</div>
+              :
+              <div className="btn cb-btn submit-btn cb-complete-btn" onClick={()=>completeReading(book.id, day, book)}>MARK COMPLETE</div>
+            }
           </div>
       }
       <div className="cb-column cb-btn-col">
-        <div className="btn btn-red cb-btn" onClick={()=>handleDeleteBook(book.id)}>Delete Book</div>
+        <div className="btn btn-red cb-btn" onClick={()=>setDeleteBookMode(true)}>Delete Book</div>
         <div className="btn cb-btn submit-btn" >View/Edit</div>
       </div>
       
-      
+      {/* Delete Modal */}
+      <div className={modalClassNames}>
+        <h2 className="red-text">Delete '{bookInfo.title}'?</h2>
+        <div className="deleteModal-row2">
+          <img className="" src={bookInfo.thumbnail} alt="book cover"/>
+          <div className="deleteModal-row2-col2">
+            <p className="deleteModal-para">Are you sure you want to delete '{bookInfo.title}'?</p>
+            <br/>
+            <p className="deleteModal-para">Once you delete a title, the book and any progress will be permanently deleted.  You can of course add the title back at any time.</p>
+            
+          </div>
+        </div>
+        <div className="btn btn-red modal-btn" onClick={()=>handleDeleteBook()}>I understand and want to delete</div>
+        <div className="btn submit-btn modal-btn" onClick={()=>setDeleteBookMode(false)}>Take me back!</div>
+      </div>
       
     </div>
   )
